Improve simulation data.  But unknown problem causes crash if transition needed for proper format is included.  Still investigating.

+++ source/PXXAnalyzerResults.cpp
@@ -27,7 +27,7 @@ void PXXAnalyzerResults::GenerateBubbleText( U64 frame_index, Channel& channel, 
   if (frame.mType == 1) strcat(number_str, " (START)");
   if (frame.mType == 2) strcat(number_str, frame.mFlags ? " (CRC ERR)" : " (END)");
   AddResultString(number_str);
-  if (frame.mData2) sprintf(&number_str[strlen(number_str)], " %04X", frame.mData2);
+  if (frame.mData2) sprintf(&number_str[strlen(number_str)], " %04X", (unsigned)frame.mData2);
 	AddResultString( number_str );
 }
